Refactor support specs
Move some of the common steps to their own module.
