+++ spec/support/system/activate_features_steps.rb
@@ -1,0 +1,21 @@
+module ActivateFeaturesSteps
+  def given_the_service_is_open
+    FeatureFlags::FeatureFlag.activate(:service_open)
+  end
+
+  def and_staff_http_basic_is_active
+    FeatureFlags::FeatureFlag.activate(:staff_http_basic_auth)
+  end
+
+  def and_the_eligibility_screener_is_enabled
+    FeatureFlags::FeatureFlag.activate(:eligibility_screener)
+  end
+
+  def and_the_referral_form_feature_is_active
+    FeatureFlags::FeatureFlag.activate(:referral_form)
+  end
+
+  def and_the_eligibility_screener_feature_is_active
+    FeatureFlags::FeatureFlag.activate(:eligibility_screener)
+  end
+end
