+++ spec/system/manage/case_worker_with_basic_auth_is_not_authorized_to_view_public_referrals_spec.rb
@@ -4,17 +4,13 @@ require "rails_helper"
 RSpec.feature "Manage referrals" do
   include CommonSteps
 
-  before do
-    create_list(:referral, 30, :submitted)
-    create_list(:referral, 20)
-  end
-
   scenario "Case worker with basic auth is not authorized to view public referrals",
            type: :system do
     given_the_service_is_open
     and_the_referral_form_feature_is_active
     and_the_eligibility_screener_feature_is_active
     and_staff_http_basic_is_active
+    and_referral_forms_exist
 
     when_i_am_authorized_with_basic_auth_as_a_case_worker
     when_i_visit_the_referrals_page
