+++ spec/system/manage/case_worker_without_permissions_is_not_authorized_to_view_public_referrals_spec.rb
@@ -4,16 +4,12 @@ require "rails_helper"
 RSpec.feature "Manage referrals" do
   include CommonSteps
 
-  before do
-    create_list(:referral, 30, :submitted)
-    create_list(:referral, 20)
-  end
-
   scenario "Case worker without manage referrals permissions is not authorized to view public referrals",
            type: :system do
     given_the_service_is_open
     and_the_referral_form_feature_is_active
     and_the_eligibility_screener_feature_is_active
+    and_referral_forms_exist
 
     when_i_am_authorized_as_a_case_worker_without_management_permissions
     when_i_visit_the_referrals_page
